Added tree overlays

+++ Bermuda/Bermuda/GameTimer.cpp
@@ -17,6 +17,7 @@ void GameTimer::init()
 	this->currentDayPart = DAYPART::Day;
 
 	//Draw part
+	
 	textCircle = IMG_LoadTexture(GameStateManager::Instance()->sdlInitializer->getRenderer(), (RESOURCEPATH + "HUD\\circleTotal.png").c_str());
 	textArrow = IMG_LoadTexture(GameStateManager::Instance()->sdlInitializer->getRenderer(), (RESOURCEPATH + "HUD\\arrow.png").c_str());
 
@@ -39,7 +40,7 @@ void GameTimer::updateGameTime(double frameTime)
 	this->updateDay();
 }
 
-long GameTimer::getGameTime()
+float GameTimer::getGameTime()
 {
 	return this->gameTime;
 }
